Move guild actions to endpoint.
